requirements.txt
psycopg2-binary
pandas
SQLAlchemy
python-dotenv

pandas as pd
psycopg2
psycopg2.extras import execute_batch
os
re
